Introduce new logical file structure
