+++ packages/node/src/methods/index.ts
@@ -3,23 +3,21 @@ import {
   getInstalledAppInstancesController,
   getProposedAppInstancesController
 } from "./app-instance-operations";
-import {
-  appInstanceInstallController,
-  appInstanceProposeInstallController,
-  virtualAppInstanceProposeInstallController
-} from "./install-operations";
+import { installAppInstanceController } from "./install/controller";
 import {
   createMultisigController,
   getAllChannelAddressesController
 } from "./multisig-operations";
+import { proposeInstallVirtualAppInstanceController } from "./propose-install-virtual/controller";
+import { proposeInstallAppInstanceController } from "./propose-install/controller";
 
 export {
-  appInstanceInstallController,
-  appInstanceProposeInstallController,
+  installAppInstanceController,
+  proposeInstallAppInstanceController,
   createMultisigController,
   getAllChannelAddressesController,
   getAppInstanceStateController,
   getInstalledAppInstancesController,
   getProposedAppInstancesController,
-  virtualAppInstanceProposeInstallController
+  proposeInstallVirtualAppInstanceController
 };
