+++ packages/node/test/unit/install.spec.ts
@@ -5,7 +5,7 @@ import { instance, mock, when } from "ts-mockito";
 import { v4 as generateUUID } from "uuid";
 
 import { ERRORS } from "../../src/methods/errors";
-import { install } from "../../src/methods/install-operations";
+import { install } from "../../src/methods/install/app-instance";
 import { openStateChannel } from "../../src/methods/multisig-operations";
 import { Store } from "../../src/store";
 import { EMPTY_NETWORK } from "../integration/utils";
